move advanced test to others
